update latest codebase

+++ llava_dpo/eval/model_vqa_loader.py
@@ -80,7 +80,12 @@ def eval_model(args):
     questions = get_chunk(questions, args.num_chunks, args.chunk_idx)
     answers_file = os.path.expanduser(args.answers_file)
     os.makedirs(os.path.dirname(answers_file), exist_ok=True)
-    ans_file = open(answers_file, "w")
+    if os.path.exists(answers_file):
+        existed = len([x for x in open(answers_file, 'r')])
+        ans_file = open(answers_file, "a")
+    else:
+        existed = 0
+        ans_file = open(answers_file, "w")
 
     if 'plain' in model_name and 'finetune' not in model_name.lower() and 'mmtag' not in args.conv_mode:
         args.conv_mode = args.conv_mode + '_mmtag'
@@ -88,7 +93,11 @@ def eval_model(args):
 
     data_loader = create_data_loader(questions, args.image_folder, tokenizer, image_processor, model.config)
 
+    cnt = 0
     for (input_ids, image_tensor), line in tqdm(zip(data_loader, questions), total=len(questions)):
+        cnt += 1
+        if cnt <= existed: continue
+        
         idx = line["question_id"]
         cur_prompt = line["text"]
 
@@ -120,7 +129,7 @@ def eval_model(args):
                                    "answer_id": ans_id,
                                    "model_id": model_name,
                                    "metadata": {}}) + "\n")
-        # ans_file.flush()
+        ans_file.flush()
     ans_file.close()
 
 if __name__ == "__main__":
